chore: move interview page components to @lib/components/interview
