+++ app/interview/page.tsx
@@ -1,4 +1,4 @@
-import { InterviewInterfaceComponent } from "@/components/interview-interface";
+import { InterviewInterfaceComponent } from "@/components/interview/InterviewInterface";
 
 export default function InterviewPage() {
   return <InterviewInterfaceComponent />;
