+++ components/interview/QuestionDisplay.tsx
