+++ components/interview/QuestionHeader.tsx
@@ -1,4 +1,4 @@
-import { Timer } from "@/components/Timer";
+import { Timer } from "@/components/interview/Timer";
 import React from "react";
 
 type QuestionsHeaderProps = {
