+++ components/interview/QuestionReady.tsx
